Refactored upload approved orders test flow
