+++ playwright-e2e/test/contested/FR_uploadApprovedOrder/upload_approved_order.spec.ts
@@ -1,10 +1,11 @@
-import { expect, test } from '../../../fixtures/fixtures.ts';
+import {test} from '../../../fixtures/fixtures.ts';
 import config from '../../../config/config.ts';
-import { ContestedCaseFactory } from '../../../data-utils/factory/contested/ContestedCaseFactory.ts';
-import { CommonEvents, ContestedEvents } from '../../../config/case-data.ts';
-import { YesNoRadioEnum } from '../../../pages/helpers/enums/RadioEnums.ts';
-import { migratedUploadApprovedOrderTabDataOnHearing1, newUploadApprovedOrderMHTabDataOnHearing1 } from '../../../resources/tab_content/contested/hearings_tabs.ts';
+import {ContestedCaseFactory} from '../../../data-utils/factory/contested/ContestedCaseFactory.ts';
+import {CommonEvents} from '../../../config/case-data.ts';
+import {YesNoRadioEnum} from '../../../pages/helpers/enums/RadioEnums.ts';
+import {newUploadApprovedOrderMHTabDataOnHearing1} from '../../../resources/tab_content/contested/hearings_tabs.ts';
 import {AxeUtils} from "../../../fixtures/utils/axe-utils.ts";
+import {UploadApprovedOrderMHPage} from "../../../pages/events/upload-approved-order/UploadApprovedOrderMHPage.ts";
 
 async function loginAsCaseWorker(caseId: string, manageCaseDashboardPage: any, loginPage: any): Promise<void> {
     await manageCaseDashboardPage.visit();
@@ -12,48 +13,18 @@ async function loginAsCaseWorker(caseId: string, manageCaseDashboardPage: any, l
     await manageCaseDashboardPage.navigateToCase(caseId);
 }
 
-async function performUploadApprovedOrderFlow(
+async function performUploadApprovedOrderFlowMH(
   caseDetailsPage: any,
-  uploadApprovedOrderPage: any,
-  testInfo: any,
+  uploadApprovedOrderMHPage: UploadApprovedOrderMHPage,
   axeUtils: AxeUtils
 ): Promise<void> {
   await caseDetailsPage.selectNextStep(CommonEvents.uploadApprovedOrder);
-  await uploadApprovedOrderPage.uploadFirstUploadApprovedOrder('approvedOrder.pdf');
-  await uploadApprovedOrderPage.navigateContinue();
-  await uploadApprovedOrderPage.selectJudge('District Judge');
-  await uploadApprovedOrderPage.enterJudgeName('District Judge Smith');
-  await uploadApprovedOrderPage.enterCourtOrderDate('01', '01', '2022');
-  await uploadApprovedOrderPage.navigateContinue();
-  await uploadApprovedOrderPage.addNewAdditionalHearingDetails();
-  await uploadApprovedOrderPage.selectFirstFinalOrder(YesNoRadioEnum.YES);
-  await uploadApprovedOrderPage.selectFirstIsAnotherHearing(YesNoRadioEnum.YES);
-  await uploadApprovedOrderPage.enterFirstTimeEstimate('30 minutes');
-  await uploadApprovedOrderPage.enterFirstHearingTime('10:00');
-  await uploadApprovedOrderPage.enterHearingDate('01', '01', '2022');
-  await uploadApprovedOrderPage.selectFirstHearingType('Final Hearing (FH)');
-
-  const courtName: string = "BIRMINGHAM CIVIL AND FAMILY JUSTICE CENTRE";
-  await uploadApprovedOrderPage.selectCourtZoneDropDown(courtName);
-
-  await uploadApprovedOrderPage.navigateContinue();
-  // CYA page
-  await uploadApprovedOrderPage.navigateSubmit();
-}
-
-async function performUploadApprovedOrderFlowMH(
-  caseDetailsPage: any,
-  uploadApprovedOrderPage: any,
-  uploadApprovedOrderMHPage: any,
-  testInfo: any,
-  axeUtils: AxeUtils
-): Promise<void> {
-  await caseDetailsPage.selectNextStep(ContestedEvents.uploadApprovedOrderMH);
-  await uploadApprovedOrderPage.uploadFirstUploadApprovedOrder('approvedOrder.pdf');
+  await uploadApprovedOrderMHPage.uploadFirstUploadApprovedOrder('approvedOrder.pdf');
   await uploadApprovedOrderMHPage.navigateContinue();
-  await uploadApprovedOrderPage.selectJudge('District Judge');
-  await uploadApprovedOrderPage.enterJudgeName('District Judge Smith');
-  await uploadApprovedOrderPage.enterCourtOrderDate('01', '01', '2022');
+  await uploadApprovedOrderMHPage.selectJudge('District Judge');
+  await uploadApprovedOrderMHPage.enterJudgeName('District Judge Smith');
+  await uploadApprovedOrderMHPage.enterCourtOrderDate('01', '01', '2022');
+  await axeUtils.audit();
   await uploadApprovedOrderMHPage.navigateContinue();
   await uploadApprovedOrderMHPage.selectIsThisFinalOrder(YesNoRadioEnum.YES);
   await uploadApprovedOrderMHPage.selectDoYouWantToAddHearing(YesNoRadioEnum.YES);
@@ -67,49 +38,13 @@ async function performUploadApprovedOrderFlowMH(
   await uploadApprovedOrderMHPage.whetherToUploadOtherDocuments(YesNoRadioEnum.YES);
   await uploadApprovedOrderMHPage.uploadOtherDocuments('OtherDoc.doc');
   await uploadApprovedOrderMHPage.selectSendNoticeOfHearing(YesNoRadioEnum.YES);
-
+  await axeUtils.audit();
   await uploadApprovedOrderMHPage.navigateContinue();
   // CYA page
   await uploadApprovedOrderMHPage.navigateSubmit();
 }
 
-
-async function performManageHearingsMigration(
-  caseDetailsPage: any,
-  blankPage: any,
-  testInfo: any,
-  axeUtils: any
-): Promise<void> {
-
-  await caseDetailsPage.selectNextStep(ContestedEvents.manageHearingsMigration);
-  await axeUtils.audit();
-  await blankPage.navigateSubmit();
-  await caseDetailsPage.checkHasBeenUpdated('(Migration) Manage Hearings');
-}
-
 test.describe('Contested - Upload Approved Order (caseworker)', () => {
-  test(
-    'Form A case shows old-style hearings created by upload approved order event on the new hearing tab',
-    { tag: [] },
-    async ({
-        loginPage,
-        manageCaseDashboardPage,
-        caseDetailsPage,
-        uploadApprovedOrderPage,
-        blankPage,
-        axeUtils,
-      },
-      testInfo
-    ) => {
-      const caseId = await ContestedCaseFactory.createAndProcessFormACaseUpToIssueApplication();
-      await ContestedCaseFactory.caseWorkerProgressToGeneralApplicationOutcome(caseId);
-      await loginAsCaseWorker(caseId, manageCaseDashboardPage, loginPage);
-      await performUploadApprovedOrderFlow(caseDetailsPage, uploadApprovedOrderPage, testInfo, axeUtils);
-      await performManageHearingsMigration(caseDetailsPage, blankPage, testInfo, axeUtils);
-      await caseDetailsPage.assertTabData(migratedUploadApprovedOrderTabDataOnHearing1);
-    }
-  );
-  
   test('New Upload Approved Order (MH) event with hearing', { tag: [] }, async ({
     loginPage,
     manageCaseDashboardPage,
@@ -123,7 +58,7 @@ test.describe('Contested - Upload Approved Order (caseworker)', () => {
     const caseId = await ContestedCaseFactory.createAndProcessFormACaseUpToIssueApplication();
     await ContestedCaseFactory.caseWorkerProgressToGeneralApplicationOutcome(caseId);
     await loginAsCaseWorker(caseId, manageCaseDashboardPage, loginPage);
-    await performUploadApprovedOrderFlowMH(caseDetailsPage, uploadApprovedOrderPage, uploadApprovedOrderMHPage, testInfo, axeUtils);
+    await performUploadApprovedOrderFlowMH(caseDetailsPage, uploadApprovedOrderMHPage, axeUtils);
     await caseDetailsPage.assertTabData(newUploadApprovedOrderMHTabDataOnHearing1);
   });
 });
